Split py_graph into header and source files; config.hpp may be removed in the future
